RAG_APPLICATION/requirements.txt
streamlit>=1.28.0
sentence-transformers>=2.2.0
huggingface-hub>=0.19.0
google-generativeai>=0.3.0
python-dotenv>=1.0.0
